Split out the other major checks from BagVerifier
